refactor: 테스트 코드 도메인별로 패키지 생성
