+++ src/main/resources/mybatis/mapper/UserMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.coupang.marketplace.repository.UserRepository">
+<mapper namespace="com.coupang.marketplace.user.repository.UserRepository">
     <insert id="insertUser" parameterType="User">
         INSERT INTO USER (name, email, salt, password, phone) VALUES (#{name}, #{email}, #{salt}, #{password}, #{phone})
     </insert>
