+++ src/test/java/com/coupang/marketplace/fixture/UserFixture.java
@@ -1,6 +1,6 @@
 package com.coupang.marketplace.fixture;
 
-import com.coupang.marketplace.domain.User;
+import com.coupang.marketplace.user.domain.User;
 
 public class UserFixture {
